+++ packages/react-ai/src/hooks/shared.ts
@@ -1,7 +1,9 @@
-import { DataState } from '@aws-amplify/ui-react-core';
 import { GraphQLFormattedError } from '../types';
 
-export type DataClientState<T> = Omit<DataState<T>, 'message'> & {
+export interface AiClientState<T> {
+  data: T;
+  hasError: boolean;
+  isLoading: boolean;
   /**
    * @deprecated will be removed in a future major version. Superseded by `errors`
    * @description errors returned from the websocket connection
@@ -11,12 +13,12 @@ export type DataClientState<T> = Omit<DataState<T>, 'message'> & {
    * @description errors returned from the websocket connection
    */
   errors?: GraphQLFormattedError[];
-};
+}
 
-export type DataClientResponse<T> = {
+export interface AiClientResponse<T> {
   data: T | null;
   errors?: GraphQLFormattedError[];
-};
+}
 
 // default state
 export const INITIAL_STATE = {