+++ packages/react-storage/src/components/StorageBrowser/createStorageBrowser/createStorageBrowser.tsx
@@ -4,7 +4,6 @@ import { setUserAgent } from '@aws-amplify/ui';
 
 import { VERSION } from '../../../version';
 
-import { ExtendedActionConfigs } from '../actions';
 import { ErrorBoundary as DefaultErrorBoundary } from '../ErrorBoundary';
 import { useAction } from '../useAction';
 import { assertRegisterAuthListener } from '../validators';
@@ -43,11 +42,9 @@ const UA_CONFIG = {
  * @returns `StorageBrowser` component, `useAction` and `useView` hooks
  */
 export default function createStorageBrowser<
-  Input extends CreateStorageBrowserInput,
-  RInput extends Input['actions'] extends ExtendedActionConfigs
-    ? Input['actions']
-    : ExtendedActionConfigs,
->(input: Input): CreateStorageBrowserOutput<RInput> {
+  TInput extends CreateStorageBrowserInput,
+  TActions extends NonNullable<TInput['actions']>,
+>(input: TInput): CreateStorageBrowserOutput<TActions> {
   assertRegisterAuthListener(input.config.registerAuthListener);
 
   setUserAgent(UA_CONFIG);
@@ -64,8 +61,8 @@ export default function createStorageBrowser<
       : input.ErrorBoundary ?? DefaultErrorBoundary;
 
   const StorageBrowser: StorageBrowserType<
-    DerivedActionViewType<RInput>,
-    DerivedActionViews<RInput>
+    DerivedActionViewType<TActions>,
+    DerivedActionViews<TActions>
   > = (props) => (
     <ErrorBoundary>
       <Provider {...props}>
@@ -75,7 +72,9 @@ export default function createStorageBrowser<
   );
 
   StorageBrowser.LocationActionView =
-    LocationActionView as LocationActionViewType<DerivedActionViewType<RInput>>;
+    LocationActionView as LocationActionViewType<
+      DerivedActionViewType<TActions>
+    >;
   StorageBrowser.LocationDetailView = LocationDetailView;
   StorageBrowser.LocationsView = LocationsView;