+++ packages/react-storage/src/components/StorageBrowser/useAction/types.ts
@@ -1,9 +1,8 @@
 import React from 'react';
-import { DataState } from '@aws-amplify/ui-react-core';
+import { AsyncReducerState } from '@aws-amplify/ui-react-core';
 
 import {
   ActionHandler,
-  ExtendedActionConfigs,
   CopyHandler,
   CreateFolderHandler,
   DeleteHandler,
@@ -14,9 +13,10 @@ import {
   UploadHandler,
 } from '../actions';
 import { ProcessTasksOptions, StatusCounts, Task } from '../tasks';
+import { StorageBrowserActions } from '../createStorageBrowser';
 
 export type ListActionState<T = any, K = any> = [
-  state: DataState<T>,
+  state: AsyncReducerState<T>,
   handleAction: (...input: K[]) => void,
 ];
 
@@ -41,18 +41,16 @@ export interface ActionHandlersProviderProps extends ActionHandlersContext {
   children?: React.ReactNode;
 }
 
-type DerivedCustomActions<T> = T extends { custom?: infer U } ? U : {};
-
 export type ResolveHandlerType<T> = T extends { handler: infer X } | infer X
   ? X
   : never;
 
-export type DerivedActionHandlers<
-  C extends ExtendedActionConfigs = ExtendedActionConfigs,
-  D extends DerivedCustomActions<C> = DerivedCustomActions<C>,
-> = DefaultActionHandlers & {
-  [K in keyof D]: ResolveHandlerType<D[K]>;
-};
+export type DerivedActionHandlers<TActions extends StorageBrowserActions> =
+  DefaultActionHandlers & {
+    [K in keyof NonNullable<TActions['custom']>]: ResolveHandlerType<
+      NonNullable<TActions['custom']>[K]
+    >;
+  };
 
 export type InferTask<THandler> = THandler extends ActionHandler<
   infer TData,